// src/app/models/category.model.ts
export interface Category {
    key: string;
    name: string;
    parentId: number;
    thumb: string;
    altname: string;
}
// export interface Category {
//     key: string;
//     name: string;
//     thumb: string;
//     subcat: SubCategory[];
// }

// export interface SubCategory {
//     key: string;
//     name: string;
//     thumb: string;
// }